Increase number of steps in Blanks

+++ crates/ppu/src/cycle.rs
@@ -1,5 +1,5 @@
 #[derive(Debug, PartialEq, Eq, Clone, Copy)]
 pub enum Cycle {
-    Ppu(u8),
+    Ppu(u32),
     Finished,
 }
